abdominal-trauma-detection/requirements.txt
torch==2.0.0
torchvision==0.15.1
monai==1.3.0
opencv-python==4.8.0.76
numpy==1.24.3
pandas==2.0.3
tqdm==4.65.0
scikit-learn==1.3.0
ultralytics==8.0.147
pydicom==2.4.3
matplotlib==3.7.2
pyyaml==6.0
scipy==1.11.1
Pillow==10.0.0
requests==2.31.0
tensorboard==2.13.0
jupyter==1.0.0
ipython==8.15.0
kaggle==1.5.13
